requirements.txt
Django>=4.0, <5.0
psycopg2>=2.8
django-environ
djangorestframework
markdown
django-filter
requests